[Gold III] Title: 투스타 춘배, Time: 76 ms, Memory: 12872 KB -BaekjoonHub

+++ 백준/Gold/30414. 투스타 춘배/투스타 춘배.cc
@@ -8,15 +8,15 @@ int N, P;
 vector<int> tree[100005];
 int A[100005], B[100005];
 
-ll DFS(int node, int par) {
-    ll ret = A[node] - B[node];
+int DFS(int node, int par) {
+    int ret = A[node] - B[node];
 
     for (int next: tree[node]) {
         if (next == par) continue;
         ret += DFS(next, node);
     }
 
-    return min(0LL, ret);
+    return min(0, ret);
 }
 
 void solve() {
